requirements.txt
langchain
langchain-core
langchain-community
pypdf
pymupdf
tqdm
sentence-transformers
faiss-cpu
chromadb
langchain-groq
python-dotenv
langchain-huggingface
langchain-ollama
langgraph
fastapi
uvicorn
langchain-google-genai
langchain-chroma